requirements.txt
configparser==3.5.0
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
mysqlclient==1.3.12
Pillow==5.3.0
protobuf==3.6.1
pytz==2018.5
six==1.11.0
